2026-03-23 v11.0.1: IPLayerCapacity Speed Test
Added
- IPLayerCapacity() Speed Test command (which uses OB-UDP-ST). To
enable, comment out REMOVE_IP_CAPACITY_DIAG in vendor_defs.h

Fixed
- Multiple EndpointID or Password TLVs in UDS handshake frame should be
handled as protocol error
- USP Service is prevented from re-registering DM if it disconnects
while the USP Broker is waiting for a response from it
- Build failure in database.c on some platforms due to compiler type
size issues

+++ src/core/database.c
@@ -965,7 +965,7 @@ int OpenUspDatabase(char *db_file)
     #define SQLITE_MINIMUM_PAGE_SIZE 512
     if (st.st_size < SQLITE_MINIMUM_PAGE_SIZE)
     {
-        USP_LOG_Error("%s: USP database (%s) size (%ld) is less than the sqlite minimum page size (%d)", __FUNCTION__, db_file, st.st_size, SQLITE_MINIMUM_PAGE_SIZE);
+        USP_LOG_Error("%s: USP database (%s) size (%ld) is less than the sqlite minimum page size (%d)", __FUNCTION__, db_file, (long)st.st_size, SQLITE_MINIMUM_PAGE_SIZE);
         err = ReInitializeDatabase(db_file, true);
         goto exit;
     }
